layout/forms/test/test_unstyled_control_height.html

<!doctype html>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<style>
  .non-native * {
    appearance: none;
  }
  .no-padding * {
    padding: 0;
  }
</style>
<div id="container" class="non-native">
  <input>
  <button>Foo</button>
  <select><option>Foo</option></select>
</div>
<script>
function testHeightMatches(desc) {
  let commonHeight = null;
  for (let element of document.querySelectorAll("#container > *")) {
    let height = element.getBoundingClientRect().height;
    if (commonHeight === null) {
      commonHeight = height;
    }
    is(height, commonHeight, "Height of the controls should match" + desc);
  }
}

testHeightMatches("");

document.getElementById("container").classList.add("no-padding");

testHeightMatches(" without padding");
</script>
